refactor: replace footer buttons with reusable SearchFilterPageFooterButtons component across search filters
